[Globals] Move Actor struct into separate file
