+++ src/globals/Actor.cpp
@@ -1,0 +1,3 @@
+#include "globals/Actor.h"
+
+// no-op
