+++ src/globals/Globals.h
@@ -69,20 +69,6 @@ enum class MainWidgets
     Downloads
 };
 
-struct Actor
-{
-    QString name;
-    QString role;
-    QString thumb;
-    QByteArray image;
-    QString id;
-    int order = 0; // used by Kodi NFO
-    bool imageHasChanged = false;
-};
-Q_DECLARE_METATYPE(Actor*)
-Q_DECLARE_METATYPE(QString*)
-Q_DECLARE_METATYPE(QVector<int>)
-
 struct DiscographyAlbum
 {
     QString title;
